Serve static files from fustily app
#21
