+++ package.json
@@ -25,6 +25,7 @@
     "envalid": "6.0.0",
     "fastify": "2.11.0",
     "fastify-cors": "3.0.1",
+    "fastify-static": "^2.6.0",
     "fetch-resolver": "1.3.0",
     "fetchq": "2.6.2",
     "react": "16.12.0",
