+++ ssr/service/service-fastify-static/hooks.js
@@ -1,0 +1,7 @@
+const { SERVICE } = require('@forrestjs/hooks');
+
+const SERVICE_NAME = `${SERVICE} fastify-static`;
+
+module.exports = {
+  SERVICE_NAME,
+};
