+++ ssr/service/service-fastify-static/index.js
@@ -1,9 +1,9 @@
-const corsPlugin = require('fastify-cors');
+const staticPlugin = require('fastify-static');
 const hooks = require('./hooks');
 
 const onFastifyHacksBefore = ({ registerPlugin }, { getConfig }) => {
-  const options = getConfig('fastify.cors', {});
-  registerPlugin(corsPlugin, options);
+  const options = getConfig('fastify.static', {});
+  registerPlugin(staticPlugin, options);
 };
 
 module.exports = ({ registerAction }) => {
